requirements.txt
blinker==1.9.0
branca==0.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
folium==0.19.4
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
googlemaps==4.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.2.2
osmnx==2.0.1
packaging==24.2
pandas==2.2.3
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
xyzservices==2025.1.0
